telegram-bot/requirements.txt
aiogram==3.17.0
python-dotenv==1.0.1
httpx==0.28.1
